feat(browser-emulator): migrate away from proxies to custom descriptors
Using proxies has a huge amount of flaws, and has proven to be very difficult to fix.
In some cases even possible to fix (proxy != transparent wrapper).
But we dont need the full power of proxies we only need to be able to modify functions.
This commit migrates most proxy logic to the new internalModifyDescriptor logic.
And fixes problems due to TS strict mode